pyproject.toml
[project]
name = "diet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "python-telegram-bot[job-queue,webhooks]>=21.6",
    "google-generativeai>=0.8.3",
    "pydantic>=2.9.2",
    "uv>=0.5.2",
    "sync>=1.0.0",
    "aider-chat>=0.63.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
]
